Prepend c.EventsAfter to msg.Chunk
This resolves an issue with two message where it wouldn't find the
command event located in c.EventsAfter, causing it to loop infinitely
and delete future messages and placing extraneous load on the
homeserver.

This could potentially be improved in the future such as by
preallocating a slice with a capacity of msg.Chunk+1 but this is
currently fine as-is.